+++ mlir/lib/IR/PDL/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_mlir_library(MLIRIRPDLPatternMatch
+  PDLPatternMatch.cpp
+
+  ADDITIONAL_HEADER_DIRS
+  ${MLIR_MAIN_INCLUDE_DIR}/mlir/IR
+)
+
