+++ mlir/lib/Rewrite/CMakeLists.txt
@@ -1,5 +1,6 @@
+set(LLVM_OPTIONAL_SOURCES ByteCode.cpp)
+
 add_mlir_library(MLIRRewrite
-  ByteCode.cpp
   FrozenRewritePatternSet.cpp
   PatternApplicator.cpp
 
@@ -11,8 +12,31 @@ add_mlir_library(MLIRRewrite
 
   LINK_LIBS PUBLIC
   MLIRIR
-  MLIRPDLDialect
-  MLIRPDLInterpDialect
-  MLIRPDLToPDLInterp
   MLIRSideEffectInterfaces
   )
+
+if(MLIR_ENABLE_PDL_IN_PATTERNMATCH)
+  add_mlir_library(MLIRRewritePDL
+    ByteCode.cpp
+
+    ADDITIONAL_HEADER_DIRS
+    ${MLIR_MAIN_INCLUDE_DIR}/mlir/Rewrite
+
+    DEPENDS
+    mlir-generic-headers
+
+    LINK_LIBS PUBLIC
+    MLIRIR
+    MLIRPDLDialect
+    MLIRPDLInterpDialect
+    MLIRPDLToPDLInterp
+    MLIRSideEffectInterfaces
+  )
+
+  target_link_libraries(MLIRRewrite PUBLIC
+    MLIRPDLDialect
+    MLIRPDLInterpDialect
+    MLIRPDLToPDLInterp
+    MLIRRewritePDL)
+endif()
+
